feat(as-sdk): throw on invalid JSON input
Update "json-as" to v0.9.26

There are still some values which should be invalid that get accepted,
such as a string without the closing double quotes, but overall the
error message for the edge cases we had have improved significantly.

Stack traces are still poor due to the constraints of WASM, but there is
work being done on a transform to allow try/catch and Rust style
results.