feat(dev-tools): support signer with multiple addresses
While it's a little inefficient to create the same wallet multiple times
for the different addresses, it does  mean we don't have to worry about
all the existing code that takes a signer as an argument. Given that the
single address flow is going to be the most common by far it makes sense
to have this pattern as leading in the interfaces.